EC-838 Renaming JobsService
