+++ NICE.Identity.Authorisation.WebAPI/Services/JobsService.cs
@@ -1,51 +1,95 @@
-using Auth0.ManagementApi;
-using Auth0.ManagementApi.Models;
-using Microsoft.Extensions.Logging;
-using NICE.Identity.Authorisation.WebAPI.Configuration;
+﻿using Microsoft.Extensions.Logging;
+using NICE.Identity.Authorisation.WebAPI.ApiModels;
+using NICE.Identity.Authorisation.WebAPI.Repositories;
 using System;
-using System.Threading.Tasks;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace NICE.Identity.Authorisation.WebAPI.Services
 {
     public interface IJobsService
     {
-        Task<Job> VerificationEmail(string authenticationProviderUserId);
+        Job CreateJob(Job job);
+        List<Job> GetJobs();
+        Job GetJob(int jobId);
+        Job UpdateJob(int jobId, Job job);
+        int DeleteJob(int jobId);
     }
 
     public class JobsService : IJobsService
     {
-
+        private readonly IdentityContext _context;
         private readonly ILogger<JobsService> _logger;
-        private readonly IProviderManagementService _providerManagementService;
-
-        public JobsService(ILogger<JobsService> logger, IProviderManagementService providerManagementService)
+        public JobsService(IdentityContext context, ILogger<JobsService> logger)
         {
-	        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-	        _providerManagementService = providerManagementService;
+            _context = context ?? throw new ArgumentNullException(nameof(context));
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
         }
 
-        public async Task<Job> VerificationEmail(string authenticationProviderUserId)
+        public Job CreateJob(Job job)
         {
-			if (string.IsNullOrWhiteSpace(authenticationProviderUserId))
-				throw new ArgumentNullException(nameof(authenticationProviderUserId));
-
-            var managementApiAccessToken = await _providerManagementService.GetAccessTokenForManagementAPI();
-            var managementApiClient = new ManagementApiClient(managementApiAccessToken, AppSettings.ManagementAPI.Domain);
-
             try
             {
-                var sendVerificationEmail = await managementApiClient.Jobs
-                    .SendVerificationEmailAsync(new VerifyEmailJobRequest {
-                        UserId = authenticationProviderUserId
-                    });
-
-                return sendVerificationEmail;
+                var jobToCreate = new DataModels.Job();
+                jobToCreate.UpdateFromApiModel(job);
+                var createdJob = _context.Jobs.Add(jobToCreate);
+                _context.SaveChanges();
+                return new Job(createdJob.Entity);
             }
             catch (Exception e)
             {
-                _logger.LogError(e.Message);
-                throw new Exception("Error when calling the Management API.", e); 
+                _logger.LogError($"Failed to create job for {job.JobId.ToString()} - exception: {e.Message}");
+                throw new Exception($"Failed to create job for {job.JobId.ToString()} - exception: {e.Message}");
+            }
+
+         }
+
+        public List<Job> GetJobs()
+        {
+            return _context.Jobs.Select(job => new Job(job)).ToList();
+        }
+
+        public Job GetJob(int jobId)
+        {
+            var job = _context.Jobs.Where((j => j.JobId == jobId)).FirstOrDefault();
+            return job != null ? new Job(job) : null;
+        }
+
+        public Job UpdateJob(int jobId, Job job)
+        {
+            try
+            {
+                var jobToUpdate = _context.Jobs.Find(jobId);
+                if (jobToUpdate == null)
+                    throw new Exception($"Job not found {jobId.ToString()}");
+
+                jobToUpdate.UpdateFromApiModel(job);
+                _context.SaveChanges();
+                return new Job(jobToUpdate);
+            }
+            catch (Exception e)
+            {
+                _logger.LogError($"Failed to update job {jobId.ToString()} - exception: {e.Message}");
+                throw new Exception($"Failed to update job {jobId.ToString()} - exception: {e.Message}");
+            }
+        }
+
+        public int DeleteJob(int jobId)
+        {
+            try
+            {
+                var jobToDelete = _context.Jobs.Find(jobId);
+                if (jobToDelete == null)
+                    return 0;
+                _context.Jobs.RemoveRange(jobToDelete);
+                return _context.SaveChanges();
+            }
+            catch (Exception e)
+            {
+                _logger.LogError($"Failed to delete job {jobId.ToString()} - exception: {e.Message}");
+                throw new Exception($"Failed to delete job {jobId.ToString()} - exception: {e.Message}");
             }
         }
     }
+
 }
