+++ NICE.Identity.Authorisation.WebAPI/Startup.cs
@@ -50,7 +50,7 @@ namespace NICE.Identity.Authorisation.WebAPI
 			//services.TryAddSingleton<ISeriLogger, SeriLogger>();
 			services.AddTransient<IClaimsService, ClaimsService>();
 			services.AddTransient<IUsersService, UsersService>();
-            services.AddTransient<IJobsService, JobsService>();
+            services.AddTransient<IVerificationEmailService, VerificationEmailService>();
             services.AddTransient<IWebsitesService, WebsitesService>();
             services.AddTransient<IServicesService, ServicesService>();
             services.AddTransient<IEnvironmentsService, EnvironmentsService>();
@@ -59,7 +59,7 @@ namespace NICE.Identity.Authorisation.WebAPI
             services.AddTransient<IProviderManagementService, Auth0ManagementService>();
 			services.AddTransient<IOrganisationsService, OrganisationsService>();
 			services.AddTransient<IOrganisationRolesService, OrganisationRolesService>();
-			services.AddTransient<IUserJobsService, UserJobsService>();
+			services.AddTransient<IJobsService, JobsService>();
 			services.AddHttpClient(); //this adds http client factory for use in DI
 
 			services.AddRouting(options => options.LowercaseUrls = true);
